Add redux to safely store creds + theming
This commit has redux introduced to globally manage the state of a users
login but its contained within the memory of the application, reducing
potential attack surfaces. This is needed especially as Strapi has the
JWT valid for 1 month.

The custom theme was also introduced. It has generated values in for
now, but will be updated with the sites theme when that is finalised. It
is early implementation for easier changes later on.