Add navivation bar and make routing dynamic
The navigation bar and site routing are now both inferred from the same
location in the code to make implementing new pages seemless. All a dev
has to do is enter a new object in the routes table, specify if its
protected and needs to be rendered in the site navigation.