Make books page call to CMS endpoint
This is the first testing of the books endpoint. The data has been
mapped so react can easily parse and render it. Might want to come back
to this and create a mapper to store the data in an easier to access
object. This current object is messy.

Need to build a component to house all the books.